wagmi: renamed Wagmi related packages `-v2` + addressed PR comments
