+++ examples/wagmi-v2/.gitignore
