+++ examples/wagmi-v2/index.html
