+++ examples/wagmi-v2/src/main.tsx
