+++ examples/wagmi-v2/src/vite-env.d.ts
