+++ examples/wagmi-v2/tsconfig.json
