+++ examples/wagmi-v2/tsconfig.node.json
