+++ examples/wagmi-v2/vite.config.ts
