--- examples/wagmi/tests/e2e/.eslintrc.js
@@ -1,9 +1,0 @@
-const path = require("path");
-const synpressPath = path.join(
-	process.cwd(),
-	"/node_modules/@synthetixio/synpress",
-);
-
-module.exports = {
-	extends: `${synpressPath}/.eslintrc.js`,
-};
