--- examples/wagmi/tests/e2e/support.js
@@ -1,1 +1,0 @@
-import "@synthetixio/synpress/support/index";
