--- examples/wagmi/tests/e2e/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": true,
-		"baseUrl": "./",
-		"types": [
-			"cypress",
-			"@synthetixio/synpress/support",
-			"cypress-wait-until",
-			"@testing-library/cypress"
-		],
-		"outDir": "./output"
-	},
-	"include": ["**/*.*"]
-}
