+++ integrations/wagmi-v2/package.json
@@ -1,0 +1,61 @@
+{
+	"type": "module",
+	"name": "@oasisprotocol/sapphire-wagmi-v2",
+	"license": "Apache-2.0",
+	"version": "2.0.0",
+	"description": "Wagmi & Viem support for the Oasis Sapphire ParaTime.",
+	"homepage": "https://github.com/oasisprotocol/sapphire-paratime/tree/main/integrations/wagmi-v2",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/oasisprotocol/sapphire-paratime.git"
+	},
+	"keywords": ["sapphire", "paratime", "oasis", "web3", "viem", "wagmi"],
+	"files": ["dist", "!dist/*.tsbuildinfo", "src"],
+	"sideEffects": false,
+	"main": "./dist/_cjs/index.js",
+	"module": "./dist/_esm/index.js",
+	"types": "./dist/_types/index.d.ts",
+	"typings": "./dist/_types/index.d.ts",
+	"exports": {
+		"node": {
+			"import": "./dist/_esm/index.js",
+			"require": "./dist/_cjs/index.cjs",
+			"types": "./dist/_types/index.d.ts"
+		},
+		"types": "./dist/_types/index.d.ts",
+		"default": "./dist/_esm/index.js"
+	},
+	"scripts": {
+		"lint": "biome check .",
+		"format": "biome format --write .",
+		"build": "npm run clean && npm run build:cjs && npm run build:esm && npm run build:types",
+		"build:cjs": "tsc --project ./tsconfig.build.json --module commonjs --outDir ./dist/_cjs --removeComments --verbatimModuleSyntax false && printf '{\"type\":\"commonjs\"}' > ./dist/_cjs/package.json",
+		"build:esm": "tsc --project ./tsconfig.build.json --module es2015 --outDir ./dist/_esm && printf '{\"type\": \"module\",\"sideEffects\":false}' > ./dist/_esm/package.json",
+		"build:types": "tsc --project ./tsconfig.build.json --module esnext --declarationDir ./dist/_types --emitDeclarationOnly --declaration --declarationMap",
+		"clean": "rm -rf dist",
+		"prepublishOnly": "npm run build"
+	},
+	"dependencies": {
+		"@oasisprotocol/sapphire-paratime": "workspace:^",
+		"@wagmi/core": "2.x",
+		"viem": "2.x"
+	},
+	"peerDependencies": {
+		"typescript": ">=5.0.4",
+		"@wagmi/core": "2.x",
+		"viem": "2.x"
+	},
+	"peerDependenciesMeta": {
+		"typescript": {
+			"optional": true
+		}
+	},
+	"devDependencies": {
+		"@types/node": "^18.7.18",
+		"@typescript-eslint/eslint-plugin": "^5.38.0",
+		"@typescript-eslint/parser": "^5.38.0",
+		"ts-node": "^10.9.2",
+		"typescript": "^5.3.3",
+		"@biomejs/biome": "^1.6.1"
+	}
+}
