+++ integrations/wagmi-v2/tsconfig.base.json
@@ -1,0 +1,31 @@
+{
+	"include": [],
+	"compilerOptions": {
+		"incremental": false,
+
+		"strict": true,
+		"useDefineForClassFields": true,
+		"noFallthroughCasesInSwitch": true,
+		"noImplicitReturns": true,
+		"useUnknownInCatchVariables": true,
+		"noImplicitOverride": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+
+		"allowJs": false,
+		"checkJs": false,
+
+		"esModuleInterop": false,
+		"allowSyntheticDefaultImports": false,
+		"forceConsistentCasingInFileNames": true,
+		"verbatimModuleSyntax": true,
+		"importHelpers": true,
+
+		"moduleResolution": "NodeNext",
+		"module": "NodeNext",
+		"target": "ES2021",
+		"lib": ["ES2022", "DOM"],
+
+		"skipLibCheck": true
+	}
+}
