+++ integrations/wagmi-v2/tsconfig.build.json
@@ -1,0 +1,11 @@
+{
+	"extends": "./tsconfig.base.json",
+	"include": ["src"],
+	"exclude": ["src/**/*.test.ts", "src/**/*.test-d.ts", "src/**/*.bench.ts"],
+	"compilerOptions": {
+		"moduleResolution": "node",
+		"sourceMap": true,
+		"rootDir": "./src",
+		"outDir": "./dist"
+	}
+}
