+++ integrations/wagmi-v2/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	// This configuration is used for local development and type checking.
+	"extends": "./tsconfig.base.json",
+	"include": ["src"],
+	"exclude": [],
+	"compilerOptions": {
+		"baseUrl": "."
+	}
+}
