--- integrations/wagmi/src/index.ts
@@ -1,81 +1,0 @@
-import { wrapEIP1193Provider, wrap, type SapphireAnnex } from '@oasisprotocol/sapphire-paratime';
-import type { Transport } from '@wagmi/core';
-import { injected, custom } from '@wagmi/core';
-import type { EIP1193Provider } from 'viem';
-
-type Window = {
-  ethereum?: EIP1193Provider;
-};
-
-const cachedProviders: Map<EIP1193Provider, EIP1193Provider> = new Map();
-
-/**
- * Wrap the `window.ethereum` with the Sapphire encryption layer.
- * Used to provide encrypted transactions and calldata to Wagmi browser dApps.
- *
- * Example:
- * ```
- *
- *    import { injectedWithSapphire } from '@oasisprotocol/sapphire-wagmi';
- *
- *    export const config = createConfig({
- *      connectors: [
- *        injectedWithSapphire()
- *      ],
- *      ...
- *    });
- *
- * ```
- *
- * @returns Same as injected()
- */
-export function injectedWithSapphire(): ReturnType<typeof injected> {
-  return injected({
-    target: () => {
-      return {
-        id: 'injected-sapphire',
-        name: 'Injected (Sapphire)',
-        provider(window?: Window) {
-          if (window?.ethereum) {
-            // Note: providers are cached as connectors are frequently retrieved
-            //       it prevents sapphire wrapper being called multiple times
-            //       which spams the RPC with oasis_callDataPublicKey requests
-            if( ! (cachedProviders.has(window.ethereum)) ) {
-              const wp = wrapEIP1193Provider(window.ethereum as any) as EIP1193Provider & SapphireAnnex;
-              cachedProviders.set(window.ethereum, wp);
-              return wp;
-            }
-            return cachedProviders.get(window.ethereum);
-          }
-          return undefined;
-        },
-      };
-    },
-  });
-}
-
-/**
- * Provide a Sapphire encrypted RPC transport for Wagmi or Viem.
- *
- * Example:
- * ```
- *
- *    import { sapphireTransport } from '@oasisprotocol/sapphire-wagmi';
- *
- *    export const config = createConfig({
- *      transports: {
- *        [sapphireTestnet.id]: sapphireTransport()
- *      },
- *      ...
- *    });
- *
- * ```
- *
- * @returns Same as custom()
- */
-export function sapphireTransport(): Transport {
-  return (params) => {
-    const p = wrap(params.chain!.rpcUrls.default.http[0]);
-    return custom(p)(params);
-  };
-}
