--- integrations/wagmi/tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  // This file is used to compile the for cjs and esm (see package.json build scripts). It should exclude all test files.
-  "extends": "./tsconfig.base.json",
-  "include": ["src"],
-  "exclude": ["src/**/*.test.ts", "src/**/*.test-d.ts", "src/**/*.bench.ts"],
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "rootDir": "./src",
-    "outDir": "./dist"
-  }
-}
