--- integrations/wagmi/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  // This configuration is used for local development and type checking.
-  "extends": "./tsconfig.base.json",
-  "include": ["src"],
-  "exclude": [],
-  "compilerOptions": {
-    "baseUrl": "."
-  }
-}
